Server/requirements.txt
numpy==1.26.4
pandas==2.2.2
opencv-python==4.10.0.84
tensorflow==2.15.0
flask==3.0.3
flask-cors==4.0.0
pillow==10.3.0
requests==2.32.3
